+++ .travis.yml
@@ -1,0 +1,21 @@
+dist: xenial
+language: python
+python:
+  - "3.7"
+
+matrix:
+  allow_failures:
+    - python: "3.8-dev"
+    - python: "nightly"
+
+install:
+  - pip install tox-travis codecov bumpversion
+
+script:
+ - tox
+
+after_success:
+ - codecov
+ - if [ "$TRAVIS_PULL_REQUEST" = "false" ] && [ "$TRAVIS_BRANCH" = "master" ]; then bumpversion --verbose minor && bash push.sh; fi
+
+cache: pip
