+++ MANIFEST
@@ -1,0 +1,6 @@
+# file GENERATED by distutils, do NOT edit
+setup.cfg
+setup.py
+TikTokApi\__init__.py
+TikTokApi\getTrending.py
+TikTokApi\tiktok.py
