--- TikTok-Api/__init__.py
@@ -1,1 +1,0 @@
-from TikTok-Api.Tiktok import TikTokapi
