--- TikTok-Api/browsermob-proxy/README.txt
@@ -1,3 +1,0 @@
-Please see README.md (in Markdown format) or view it at:
-
-https://github.com/lightbody/browsermob-proxy
