+++ TikTokApi/__init__.py
@@ -1,0 +1,1 @@
+from TikTokApi.tiktok import TikTokapi
