+++ examples/getAUsersVideos.py
@@ -1,18 +1,14 @@
 # NOTE: This python file cannot be run from the examples folder.
 # It MUST BE in the same directory as tiktok.py is located!
 # Yeah sorry, it's kind of annoying, but it's more organized this way.
-from tiktok import TikTokapi
+from TikTokApi import TikTokapi
 
 # Starts T
-api = TikTokapi()
+api = TikTokapi("browsermob-proxy/bin/browsermob-proxy")
 
 # The Number of trending TikToks you want to be displayed
 results = 10
-
-# The TikTok user's ID, can be found in the JSON from trending
-id = "7119601"
-
-trending = api.userPosts(id, results)
+trending = api.trending(results)
 
 for tiktok in trending:
     # Prints the music play URL for the tiktok in trending
