+++ requirements.txt
@@ -1,0 +1,9 @@
+browsermob-proxy>=0.8.0
+certifi>=2019.9.11
+chardet>=3.0.4
+idna>=2.8
+psutil>=5.6.3
+requests>=2.22.0
+selenium>=3.141.0
+TikTokApi>=2.1.2
+urllib3>=1.25.3
